Declare functions within main.js
