+++ main.js
@@ -4,7 +4,38 @@ const AdmZip = require('adm-zip')
 const filesize = require('filesize')
 const pathname = require('path')
 const fs = require('fs')
-const listArtifacts = require('./listArtifacts')
+const retry = require('retry');
+
+
+async function listArtifacts(client, owner, repo, runID) {
+    let artifacts = await client.actions.listWorkflowRunArtifacts({
+        owner: owner,
+        repo: repo,
+        run_id: runID,
+    })
+
+    // One artifact or all if `name` input is not specified.
+    if (name) {
+        artifacts = artifacts.data.artifacts.filter((artifact) => {
+            return artifact.name == name
+        })
+    } else {
+        artifacts = artifacts.data.artifacts
+    }
+
+    return artifacts;
+}
+
+async function retryListArtifacts(client, owner, repo, runID) {
+
+    // TODO: create backoff here using condition artifacts.length == 0
+    // Use npm module https://npmjs.com/package/backoff
+    let artifacts = await  listArtifacts(owner, repo, runID);
+
+    if (artifacts.length == 0)
+        throw new Error("no artifacts found")
+
+}
 
 async function main() {
     try {
@@ -85,7 +116,7 @@ async function main() {
 
         console.log("==> RunID:", runID)
 
-        let artifacts = await listArtifacts.listArtifacts(client, owner, repo, runID)
+        let artifacts = await listArtifacts(client, owner, repo, runID)
 
         // TODO: create backoff here using condition artifacts.length == 0
         // Use npm module https://npmjs.com/package/backoff
@@ -126,5 +157,9 @@ async function main() {
         core.setFailed(error.message)
     }
 }
+
+
+exports.listArtifacts = listArtifacts;
+exports.retryListArtifacts = retryListArtifacts;
 
 main()
